requirements.txt
# Core ML Pipeline Framework Dependencies
# This file contains the minimal production requirements
# For development dependencies, see requirements-dev.txt

# Core data processing
pandas>=1.5.0
numpy>=1.21.0
polars>=0.18.0
duckdb>=0.8.0
pyarrow>=9.0.0

# Machine learning core
scikit-learn>=1.2.0
xgboost>=1.7.0
lightgbm>=3.3.0
catboost>=1.1.0
imbalanced-learn>=0.9.0

# Model interpretability
shap>=0.41.0
lime>=0.2.0
eli5>=0.13.0

# Data connectors
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
pymysql>=1.0.0
snowflake-connector-python>=3.0.0
redis>=4.3.0

# Big data processing
pyspark>=3.3.0
dask[complete]>=2022.8.0

# Configuration and utilities
pyyaml>=6.0
hydra-core>=1.2.0
omegaconf>=2.2.0
python-dotenv>=0.20.0

# Monitoring and logging
mlflow>=2.0.0
wandb>=0.13.0
structlog>=22.0.0

# Web framework (for APIs)
fastapi>=0.85.0
uvicorn>=0.18.0
pydantic>=1.10.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
ipywidgets>=8.0.0

# Hyperparameter optimization
optuna>=3.0.0
hyperopt>=0.2.7
scikit-optimize>=0.9.0

# Cloud and deployment
kubernetes>=24.0.0
docker>=6.0.0
boto3>=1.24.0
google-cloud-storage>=2.5.0
azure-storage-blob>=12.12.0

# Security and compliance
cryptography>=37.0.0
pyjwt>=2.4.0

# Data validation
cerberus>=1.3.0
marshmallow>=3.17.0
great-expectations>=0.15.0

# Performance
joblib>=1.2.0
numba>=0.56.0

# Utilities
click>=8.1.0
tqdm>=4.64.0
requests>=2.28.0
urllib3>=1.26.0
packaging>=21.0.0
importlib-metadata>=4.0.0

# Additional ML frameworks
h2o>=3.38.0

# Specialized connectors
pyhive>=0.6.5

# Monitoring
prometheus-client>=0.16.0